requirements.txt
fastapi
uvicorn
psycopg2
python-dotenv
pillow
sqlalchemy
psycopg2-binary
jinja2
python-multipart
pytesseract
opencv-python